fix: rule-simplification is now much faster
I'm the dumbest programmer on the face of the earth. Tales will be told,
and songs will be sung, about the great Ross the Imbecilic.

The performance issues didn't really have anything to do with the issues
I wrote about in the prior commit, mostly. It's true that the recursion
and the looping cause the rules to be checked several times each, but
that appears to not be that big of a deal, as long as, and this is key,
you don't parse every single rule, on every single loop, of every single
recursion.

I was running the parser every time. All I had to do to fix the issue
was to move the parsing outside rule-simplify, and my benchmarks went
from ~1200 milliseconds, to 3. That's a 400x speed up for copy-pasting
some lines.

Let this be a lesson to you. You rarely need to worry about algorithmic
complexity (big-O, as the cool kids call it) until you truly have a
rediculous number of items, like in a database. For most finite
problems, you just need some solid benchmarks, some experimentation, and
to be marginally smarter than a seagull.