feat: compute subspans of numbers in expressions
Before compute-numbers would only work if every operand was a number,
2*3*4=24. Now it will work on sections, 2*x*4=8*x. This required that I
disable the rules to flatten multary operations, addition and
multiplication. These rules aren't working right at the moment in many
scenarios, and it appears I might have to extend the rules engine, and
possibly mafs, to accomidate them. That or just have code outside of the
rules engine like I have with computaton of numbers.