feat: replace old simplify with new rule-simplify
The only part of the old simplification that I had to keep was the part
that flattens nested associative and commutative operations,
(a+b)+c=a+b+c. I'll just have to keep this for now, as mafs can't really
describe it well enough. Other than that old simplify is gone. I did
have to update some tests to match the way the new simplification works,
and I had to improve the results of pretty-print so that it wouldn't
produce single operand expressions, but other than that it seems to be
working very well.