feat: don't expand numeric expressions too much
The function compute-numeric now checks to see if the resulting number
has a worse latex-score than the origional expression. This means that
an expression like 2^40, which is somewhere around thirteen digits long
when expanded, will just stay as 2^40. I needed to extend latex-score to
check the digits of numbers, and while testing it I found that it
couldn't accept named symbols like pi, so added that as well. Named
symbols will just have a latex score of 1, which I think is fine for
now.