feat: rule-simplify now supports numeric fractions
Normally computation of numbers happens in a recursive bottom-up
process, such that ((1+2)*3) will add 1 and 2, and then mutliply by 3.
For fractions, however, I can't follow this process, as the computation
of 1/3 would produce an imprecise floating-point number. I had to extend
actual-comp, the part of rule-simplify that handles numbers, to
understand fractions as-well. It can now add, multiple, exponentiate,
and divide fractions of numbers. This involved bringing back a lot of
code from the insta- function days. Search for `gcd` to see those
commits.