feat: added production build code
To access the production build, you must run the code with the `BUILD`
env variable set to "PROD". This will tell the discord server to run on
port :443, the https port, and to redirect port :80, the http port, to
port :443. It will also cause google sign-in to use the spacedmath.com
redirect url. I also added the database and the pem files (which are
used for ssl encrypting https) to the gitignore file.

This code has actually been running on my servers for a while, but I
haven't bothered to check it in until now.