fix: Don't break when parsing "y="
The mafs code "y=" would create a NaN as it assumed that the empty space
was a number. I hardened the code against NaN. It's a temporary fix at
best. I should solve this problem all the way in the mafs parser,
otherwise NaN output becomes a feature, and that's not good. Simply
early.