fix: app no longer mixes problem sources
Apparently I was still looking to the old math list for my list of
skills, amoung other things. I've completely ripped math-list out of the
main app and have now switched entirely over to sqlite. This fixed an
issue where the only skills showing up for filtering were ones in the
unchanging math-list, causing trig to not be included. It also fixed a
crash that I think occured whenever the sqlite list became bigger than
the math-list, creating an out of bounds exception.

The way I handled problems and available skills as separate atoms might
not be the most idiomatic, proper way to deal with two pieces of data
that are so closely coupled to each other. One of them is entirely
dependant on the other, and I really should have conjoined them as such.
Clojure even has systems for this, but at the moment app.cljs is small
and I think it will have to stand for now.