feat: replaced mafs with math.js for parsing on Eq
The eq page now uses the external library math.js for parsing the input
equations. I intend to ultimately use it to verify the input equations
by independently computing the derivative on its own and comparing that
answer to my result. This won't verify the skill detection, but if my
answer is correct, the skills are very likely correct as well. This
should speed up the process of adding equations. Parsing with math.js
also has the added benefit of less necessity for parenthesis. My BNF
grammar was very dependent on parens to properly parse the equations.
This is because interpreting math into a tree is often driven by PEMDAS,
which is very difficult to distill in an EBNF grammar. I could have
handled it semantically, but this would have been much more work, and
math.js is already well established. The inclusion of a js library does
weaken my ability to run the Clojure code in a server environment (non
CLJS), but I'm only using it as an input interpreter for now, so if I
ever did want to migrate to the server (to JVM), I could easily ditch
the math.js functionality while keeping the core of problems.clj alive.

I've renamed old-simplify to normalize-structure, and I now plan on
keeping it. The division of labor between rule-simplify and
normalize-structure is acceptable to me. Rule-simplify reduces the math
as the user sees it. Normalize-structure just irons out the way that
math is represented in-memory. This idea was inspired by the fact that
math.js likes to always nest its associative, binary operators. So 1+2+3
is interpreted as (1+2)+3. I do it differently, so I need to normalize.
That being said, I don't want Eq to simplify 2^2 to 4 (which it is
currently doing) as I want to allow derivative problems to be able to
start with these numerical expressions.

I added a default Infinite return to latex-score. I refer to this as an
Inert Value--a value of the expected return type (as opposed to nil)
which is intended to avoid any meaningful execution. I have this problem
with my real-time rendering of problems and solution in Eq, where the
system will send any intermediate input through the entire chain, no
matter how invalid. This has been breaking a lot of stuff, including
latex-score when executed with nil. Latex-score now defaults to the
highest posible score, infinite complexity, which should tell the
simplifier to ditch the new "simpler" value, without needing to build
nil handling into the system.

For now, prime-mjs and simplify-mjs are untested, but parse-mjs and
print-mjs are exhaustively tested. Since the two functions are cyclical,
I can test them both by starting with a value in either of their
domains, and sending it through one and back through the other. Luckily,
I still have math-list available, so I spun through those values and
cycled them through.