feat: Verify Eq answers with math.js
There is now a colored box on the Eq page which tells the user whether
or not math.js came up with the same answer as problems.cljs. This
equality check only works for perfectly equal in-memory structures, not
for any equivilant math. I have yet to have found an equation which
breaks it, but I am certain they exist. This will just serve as a
convenient feature for the user you just wants to input some problems.

I had to extend prime-mjs and simplify-mjs with the normalize-structure
function so that they wouldn't nest they're associative operations. The
nesting was causing the equality to fail.