feat: Add soft delete for problems on App
I needed to save the origional string that I get from the server in
problem-list so that I have something to identify the deletion. It's
important to note that this is a "soft" delete. Those records are still
there, you just have to make an uncoped query to get them. This is vital
for now, as entering problems is time consuming, and I don't want to
risk a bug completely wiping them out. The deletion command checks that
the client who asked for the deletion is also the author of the problem,
so I have some protection from other users randomly deleting my records.