chore: added todo's after manually testing the app
I just typed up and adding dozens of equations from Stewart's Calc 8th
sections 3.1, 3.2, and 3.3. The only basic derivation problems I had to
skip was one with a bug, that matched the form y=1/(x+1), and any
problem with a theta. Theta is going to take a little bit of work, as
it will have to be a named greek letter that represents a variable. Pi
is a constant, so I assumed all named symbols would be constants, but
that's not true for theta. I'll need to make theat behave like pi in
terms of the parsing and printing, but behave like x in terms of its
treatment during derivation. I also just found some general bugs and
less than satisfactory outcomes. Everything that I noticed was recorded
in some way.