feat: fix filters and improve app page
I rearranged the problem, answer, and steps on the main app page. Steps
has been moved below the answer, separated by a dividing line, and
labeled as being "for debugging." The steps aren't always perfect, in
particular with recursion on long chains of associative operations:
a+b+c. This should allow me to demo the app without staking too much
importance on the steps display, which aren't a core feature.