fix: don't move lone constants in associative operations
I intentionally wrote the tool to change 1+x into x+1, but this was
damaging my ability to track the provenance of expression during
derivation. For example, the derivative of 1-e^x was -e^x+1, which
temporarily throws me off when checking answers. I have a goal of
leaving the answers as close to the original structure as possible, such
that the amount of work a student has to do to cross check their answer
is minimized. This moves toward that goal.