feat: added partially effective chain rule
It only works with a unary outer function at the moment, and it can
enter an endless recursive loop if no other rule matches, but it does
work. I had to put it at the end of the list of rules to set it with the
lowest priority, in order to avoid the aforementioned endless recursion.
I'd like to make this a smarter system as some point—one which may
detect the proper priority of patterns by comparing them to each other
and themselves—but this is good enough for now. In order to get this
working with binary operations, I'll need to extend the pattern matcher
to detect which operand should match with the internal `[\g \x]`
pattern.